+++ esmf-semantic-aspect-meta-model/src/test/resources/samm_2_2_0/constraint-shape/org.eclipse.esmf.test/1.0.0/TestConstraintWithoutProperties.ttl
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Robert Bosch Manufacturing Solutions GmbH
+# Copyright (c) 2024 Robert Bosch Manufacturing Solutions GmbH
 #
 # See the AUTHORS file(s) distributed with this work for additional
 # information regarding authorship.
