+++ esmf-semantic-aspect-meta-model/src/test/resources/samm_2_2_0/constraint-subclass-core/org.eclipse.esmf.test/1.0.0/TestConstraintNonUniqueLangStrings.ttl
@@ -15,8 +15,10 @@
 @prefix samm-c: <urn:samm:org.eclipse.esmf.samm:characteristic:2.2.0#> .
 @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
 
-:TestConstraintNonUniqueLangStrings a samm:Constraint ;
+:TestConstraintNonUniqueLangStrings a samm-c:LengthConstraint ;
    samm:preferredName "Test Constraint"@en ;
    samm:preferredName "Test Einschränkung"@en ;
    samm:description "TestConstraint"@en ;
-   samm:description "Test Einschränkung"@en .
+   samm:description "Test Einschränkung"@en ;
+   samm-c:minValue "5"^^xsd:nonNegativeInteger ;
+   samm-c:maxValue "10"^^xsd:nonNegativeInteger .
