+++ esmf-semantic-aspect-meta-model/src/test/resources/samm_2_2_0/constraint-subclass-core/org.eclipse.esmf.test/1.0.0/TestConstraintWithEmptyProperties.ttl
@@ -15,6 +15,8 @@
 @prefix samm-c: <urn:samm:org.eclipse.esmf.samm:characteristic:2.2.0#> .
 @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
 
-:TestConstraintWithInvalidLangStrings a samm:Constraint ;
-   samm:preferredName "Test Constraint" ;
-   samm:description "TestConstraint" .
+:TestConstraintWithEmptyProperties a samm-c:LengthConstraint ;
+   samm:preferredName ""@en ;
+   samm:description ""@en ;
+   samm-c:minValue "5"^^xsd:nonNegativeInteger ;
+   samm-c:maxValue "10"^^xsd:nonNegativeInteger .
